backend/src/data_eng/market-data-example-real/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "market-data-example"
version = "0.0.1"
description = "Minimal stage-based pipeline example for Databento MBP-10 data"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.0",
  "PyYAML>=6.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
